requirements.txt
opencv-python
PyQt5
PyYAML